// src/components/types/CardType.ts
type cardType = {
  blogno: number;
  imageType: string;
  description: string;
  blogName: string;
  slug: string;
};
export default cardType;

type blogsCardDetailsTypes = {
  blogName: string;
  firstHeading: string;
  firstPara: string;
  secondHeading: string;
  secondPara: string;
  thirdHeading: string;
  thirdPara: string;
  fourthHeading: string;
  fourthPara: string;
  fifthHeading: string;
  fifthPara: string;
  image: string;
};
export type { blogsCardDetailsTypes };